failover mechanism for DNS lookup exceptions - caused with EC2 on cumulus
Before the DNS lookup would render the entire dashboard unusuable because the exception would be uncought. Now only the service that encountered the DNS failure will become unreachable and it can be deleted.

Test Plan: I just created a PHP service on EC2 cloud that triggers the DNS lookup exception, because of a DNS issue we currently face at ZIB. The service turned unreachable and I could delete it from the dashboard.

Reviewed By: claudiugh