scalaris: manager: add cloud autoplacement
When a Scalaris service instance is started, the manager will collect the
available clouds and split them into 4 cloud groups (one for each replica). If
less than 4 clouds are specified in the director.cfg, every cloud makes up one
cloud group. The manager will balance the Scalaris agents over the groups.
For the moment, the goal is to maximise availability, i.e. keep replicas as far
apart from each other as possible.

Needs the next commit (implementing the changes in the agent) to work.