scalaris: cpsclient: add auto-placement option
Adds an auto placement option to the cpsclient for the multi-datacenter
feature.

When specifying 'auto' as value for the 'cloud' option, the manager will
distribute the new nodes as evenly as possible over the cloud groups.

If an explicit cloud is given, the new node(s) will be started in the given
cloud. In this case however, the groups are not (automatically) balanced
anymore.