Generic Service, Frontend: improved the polling mechanism
The poller starts when the page loads or when an async request completes, if there is at least one script in the RUNNING state.
It stops when no scripts are in the running state or when an async request starts (we need to avoid starting a new request before the previous one completes).
The default refresh interval for the poller is 1 second.