Improvements in deploying applications using manifests
- The user can now choose the default cloud in which to deploy an
application using manifests. This is now supported in cps-tools using
the -c option of the 'cps-tools application manifest' command. In the
web fronted, this option was also added to the manifests page and is
available for both uploaded manifests and ready-made applications.
- cps-tools now waits for an application to be ready when deploying it
from a manifest. This is in line with the semantics of all the other
cps-tools commands: the command should return only after the result is
ready.
- After uploading a manifest, the web frontend redirects the user
directly to the newly started application's page.
- Fixed a race condition that caused the application's state to briefly
switch from PROLOGUE to RUNNING during the creation of the application
from a manifest. This led to the application's page stop being refreshed
before the application is ready.
- Redesigned the "deploy an application form manifest" page in the web
frontend.
- Director: removed the useless 'cloud_name' parameter from createapp()
method: the creation of an app does not need a cloud, only starting the
app (starting the app manager) does.
- Fixed a bug in the Director related to application states;
- Fixed a bug in manifest.js
 (the file upload field was referenced with the wrong name)
- Improved log messages in Director when deploying an app using
manifests