cps-tools: proper verification of the director's SSL certificate
Before this commit, all the components (managers, agents, director,
clients) were using the pyOpenSSL library for secure HTTPS
communication, with the exception of the link between the command line
client (cps-tools) and the director, which was implemented using the
standard httplib Python library. While pyOpenSSL was doing proper
verification of the certificates, however, up to Python version 2.7.9,
the standard libraries (httplib) were not doing any HTTPS certificate
verifications. More details here:

https://www.python.org/dev/peps/pep-0476/

Even with Python versions greater than 2.7.9, verification of the
certificates was not done properly in cps-tools because of a bug
introduced in commit 8007960307788a38e58b6b8f28181418975bce82, which
disabled certificate checks globally instead of only when handling
manifests.

Now, this commit changes the way in which cps-tools and the director
communicate. They are now using the same infrastructure as the other
components, using the same helper functions implemented in the ConPaaS
core (which use pyOpenSSL). Server certificates are properly checked
every time and if the certificate verification fails, an error will
be shown to the user. The user is asked for the password only when
installing the certificates (one time), and will be authenticated from
then on only by using the certificate.

The only situation in which HTTPS certificates are not verified is when
downloading resources from external servers during a deployment of an
application from a manifest, as for the ready made applications code has
to be downloaded from https://online.conpaas.eu, which fails
verification. This verification can be easily enabled again by setting
the default value ssl_check=True for the parameter of the read_from_url()
function in cpsdirector/manifest.py (or on a call-by-call basis when this
function is called).

The commit also adds the "Connection: close" header at every request.
Without sending this header, the director keeps the connection open for
a few seconds after sending the response (persistent connections in HTTP
1.1). This leads to a 5 second delay for every call to the director,
which degrades the performance substantially. This was not an issue with
the old httplib classes, as they were not reading from the connection
until it is closed, but stopped reading when receiving a complete
response. I also modified all the calls in the director to return data
only by using the build_response() helper function, in order to make
sure that the "Connection: close" header is added.

The commit also bring improvements to the logging in cps-tools. Logging
messages can be activated by running a command with the -d flag (or by
setting the "debug=true" in cps-tools.conf). All requests sent to the
director or managers are logged, including the details of the request
and the response received.