xtreemfs: OSD drain uses SSL to connect to the DIR
This commit fixes a bug that prevented a second OSD node from shutting down properly when other OSD nodes still remain. This was caused by the drain function which pushes the data to the remaining OSD nodes before the current one shuts down, preventing data loss. The problem was that the purge function was not updated to use SSL when communicating to the DIR server, so the DIR server was just closing the connection, forcing the OSD to wait and then try again (with the same result every time). This led to the ConPaaS agent to get stuck in the EPILOGUE state.

The OSD drain function now uses the OSD's certificates to communicate to the DIR and the drain operation succeeds. Parameters (certificate name, certificate location, certificate password, trusted CA certificate location & password etc.) are passed through the agent configuration file, which makes them easily configurable if needed.