managers: added checks in service managers before certain operations
This commit allows service managers to perform their own particular checks before a new node or volume is added or removed by the application manager and to allow them the opportunity to refuse the operation if they don't allow it in the current context. This allows for example to check that enough nodes of a certain type are present before removing them or, in the case of php / java services, that enough web and backend nodes are present before adding a new proxy node. The generic service can now deny adding or removing a storage volume when the application's scripts are running in a node.

All such checks that were present before the switch to the new application manager architecture are now ported. The commit also adds additional checks (for example in the case of xtreemfs and mysql services). The code is also improved by adding a custom exception class that deals with the “wrong number of nodes to remove” error in a more user-friendly way.

Some variable names were changed in the application / service managers to be in line with the naming conventions used in the rest of the code.

cps-tools was also changed to use the default values for the number of nodes for each role (when adding / removing nodes) only when the user does not specify a value for any role (or else, the user may specify the number of nodes to be added / deleted for a role and because of the defalut, cps-tools will also request to add / remove nodes for another role).