managers: redesigned the code that assembles the service node information
This commit reimplements the transformation of the node roles dict (a dictionary containing the roles and the number of nodes of each role provided when starting or adding nodes to a service) into a list containing the node information that will be provided to the cloud controller in order to create the VM and assemble an instance of the ServiceNode class. The new design manages this transformation in an unified way (whether starting a service or adding nodes, the same code is reused) and puts in place a system in which each service manager can expose its particularities by overriding one of the methods inherited from the base manager. Such methods return the list of roles of the service, the default role (the one used when no role was specified when starting the application or adding nodes), the starting list of nodes and another function that performs the transformation from roles to the list of service node information (which can be easily implemented for any role by customizing already-provided methods inherited from the base manager).

As a result, the commit fixes a bug in the mysql service (a failure when adding both a mysql and a glb node at the same time). The new framework also allows the web services (php / java) to keep the node configuration after a restart (starting and stopping the service): if the service is stopped, when started again will power up all the VMs that were running before being stopped.

To be able to keep the role assigned to a node, the Cloud class had to be updated to transport this field from the service node information to the ServiceNode instance.

The commit also changes the text associated to the default role in cps-tools and cleans up some logs in the mysql agent.