frontend: improvements of the application page
The manager instance in the application's page was rendered using an HTML code pasted directly into the javascript instead of using the PHP code for the regular instances. I changed the implementation to use the same code. This also made the appearance of this instance in line with the others.

I also restructured the application's page to a more logical structure, that follows the same patterns as other pages (service page etc.) and removed the trailing whitespace.