mysql: manager connects to the db using the root account
This commit changes the behavior of the manager to make it use the root account (instead of the user's account) when connecting to the mysql database. This fixes a bug in which the two passwords become desynchronized and access is denied. So, now, the manager and agents only use the root password to connect the database and only store the root password, not the user's password. Access for the root account from the manager IP address is granted.

The commit also renames the IP_WHITE_LIST parameter in the agent configuration with MANAGER_IP, to better reflect its content.

Some logging messages are also improved and the root password is removed from the logs.